--- file-manager/index.js
@@ -1,1 +1,0 @@
-export { FileManager } from "./manager.js";
