+++ hash-manager/hash-manager.js
@@ -7,16 +7,27 @@ export class HashManager {
     this.manager = manager
   }
 
+  commandHandler = async (commandParts) => {
+    const filePath = commandParts[1]
+    this.calculate(filePath)
+  }
+
   calculate = async (filePath) => {
-    const readStream = createReadStream(filePath)
-    let result = ''
-    readStream.on('data', (data) => {
-        result = result + data
-    })
-    readStream.on('end', () => {
-      const hash = createHash('sha256')
-      const hashString = hash.update(result).digest('hex')
-      this.manager.message(hashString)
+    this.manager.pathMaker.checkIfPathExists(filePath).then((filePath) => {
+      if (filePath) {
+        const readStream = createReadStream(filePath)
+        let result = ''
+        readStream.on('data', (data) => {
+            result = result + data
+        })
+        readStream.on('end', () => {
+          const hash = createHash('sha256')
+          const hashString = hash.update(result).digest('hex')
+          this.manager.message(hashString)
+        })
+      } else {
+        this.manager.throwError(`Path ${filePath} does not exist`)
+      }
     })
   }
 }
