--- helpers/tt.txt
@@ -1,1 +1,0 @@
-op
