+++ package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node index.js",
-    "fast-start": "node index.js -- --username=sasaafanasienka"
+    "fast-start": "npm run start -- --username=user1"
   },
   "keywords": [
     "node",
